Move around a few more coverpoints
